+++ src/ultimate_notion/schema.py
@@ -23,10 +23,11 @@ from __future__ import annotations
 from abc import ABCMeta
 from collections import Counter
 from collections.abc import Iterator
+from functools import partial
 from textwrap import dedent
-from typing import TYPE_CHECKING, Any, TypeVar, cast
+from typing import TYPE_CHECKING, Annotated, Any, TypeVar, cast
 
-from pydantic import BaseModel, ConfigDict, Field, create_model
+from pydantic import BaseModel, ConfigDict, Field, create_model, field_validator
 from tabulate import tabulate
 
 import ultimate_notion.obj_api.schema as obj_schema
@@ -257,21 +258,49 @@ class Schema(metaclass=SchemaType):
         return [prop for prop in cls.get_props() if not prop.type.readonly]
 
     @classmethod
-    def to_pydantic_model(cls, *, with_ro_props: bool = False) -> BaseModel:
+    def to_pydantic_model(cls, *, with_ro_props: bool = False) -> type[BaseModel]:
         """Return a Pydantic model of this schema for validation.
 
         This is useful for instance when writing a web API that receives data that should be validated
-        before it is passed to Ultimate Notion.
+        before it is passed to Ultimate Notion. The actual values are converted to `PropertyValue`
+        and thus `value` needs to be called to retrieve the actual Python type.
 
         If `with_ro_props` is set to `True`, read-only properties are included in the model
         """
-        # ToDo: One could also validate the categories in Select and MultiSelect using pydantic
+
+        # ToDo: Validate the categories in Select and MultiSelect using pydantic!
+        def pytype_to_prop_value(py_type: Any, *, prop_value: type[PropertyValue]) -> PropertyValue:
+            return prop_value(py_type)
+
         kwargs: dict[str, Any] = {
-            prop.attr_name: (prop.type.prop_value.py_type, Field(default=None)) for prop in cls.get_rw_props()
+            prop.attr_name: Annotated[prop.type.prop_value, Field(default=None, alias=prop.name)]
+            for prop in cls.get_rw_props()
         }
+        validators: dict[str, Any] = {
+            f'{prop.attr_name}_validator': field_validator(prop.attr_name, mode='before')(
+                partial(pytype_to_prop_value, prop_value=prop.type.prop_value)
+            )
+            for prop in cls.get_rw_props()
+        }
+
         if with_ro_props:
-            kwargs |= {prop.attr_name: (prop.type.prop_value.py_type, Field()) for prop in cls.get_ro_props()}
-        return create_model(f'{cls.db_title}Model', __config__=ConfigDict(arbitrary_types_allowed=True), **kwargs)
+            kwargs |= {
+                prop.attr_name: Annotated[prop.type.prop_value, Field(alias=prop.name)] for prop in cls.get_ro_props()
+            }
+            validators |= {
+                f'{prop.attr_name}_validator': field_validator(prop.attr_name, mode='before')(
+                    partial(pytype_to_prop_value, prop_value=prop.type.prop_value)
+                )
+                for prop in cls.get_ro_props()
+            }
+
+        model = create_model(
+            f'{cls.db_title}Model',
+            __config__=ConfigDict(arbitrary_types_allowed=True, extra='forbid'),
+            __validators__=validators,
+            **kwargs,
+        )
+        return model
 
     @classmethod
     def to_dict(cls) -> dict[str, PropertyType]: