+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - New: Setting the width ratios of page columns is supported.
 - Fix: Adding a column with `add_column(index)` at a given index is no longer off by 1.
 - Add: Updating a database schema, i.e. adding/removing properties or changing their types
+- Add: Changing database properties, e.g. name, formula of formula property, options, etc.
 
 ## Version 0.7.1, 2025-01-05