+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix: Adding a column with `add_column(index)` at a given index is no longer off by 1.
 - Add: Updating a database schema, i.e. adding/removing properties or changing their types
 - Add: Changing database properties, e.g. name, formula of formula property, options, etc.
+- Add: Added the `display_name` field of comments
 
 ## Version 0.7.1, 2025-01-05